Split AdminInterface into two interfaces
Right now, the AdminInterface is way too big for anyone considering
implementing it. It should be split into smaller interfaces.
